refactor: merge useCodeThemes to context
